Bugfix: fixed incorrect integer definition of parameters in hysteresis loop simulations.

+++ src/program/static_hysteresis.cpp
@@ -82,11 +82,11 @@ int static_hysteresis(){
 	sim::hamiltonian_simulation_flags[3] = 0;	// Thermal
 
 	// Setup min and max fields and increment (uT)
-	int iHmax=vmath::iround(double(sim::Hmax)*1.0E6);
-	int miHmax=-iHmax;
-	int parity_old;
-	int iH_old;
-	int start_time;
+	int64_t iHmax=vmath::iround(double(sim::Hmax)*1.0E6);
+	int64_t miHmax=-iHmax;
+	int64_t parity_old;
+	int64_t iH_old;
+	int64_t start_time;
 
 	// Equilibrate system in saturation field, i.e. the largest between equilibration and maximum field set by the user
    if(sim::Heq >= sim::Hmax){
@@ -101,10 +101,10 @@ int static_hysteresis(){
 	else sim::integrate(sim::equilibration_time);
 
    // Hinc must be positive
-	int iHinc=vmath::iround(double(fabs(sim::Hinc))*1.0E6);
+	int64_t iHinc=vmath::iround(double(fabs(sim::Hinc))*1.0E6);
 
-   int Hfield;
-   int iparity=sim::parity;
+   int64_t Hfield;
+   int64_t iparity=sim::parity;
 	parity_old=iparity;
 
    // Save value of iH from previous simulation
@@ -116,7 +116,7 @@ int static_hysteresis(){
 		if(sim::load_checkpoint_flag && sim::load_checkpoint_continue_flag)
       {
          //necessary to upload value of iH_old when loading the checkpoint !!!
-		   iH_old=int(sim::iH);
+		   iH_old=static_cast<int64_t>(sim::iH); //  int(sim::iH);
          //Setup min and max fields and increment (uT)
 			if(parity_old<0){
 				if(iparity<0) miHmax=iH_old;
@@ -158,7 +158,7 @@ int static_hysteresis(){
 
 			// Increment of iH
 			Hfield+=iHinc;
-			sim::iH=int64_t(Hfield); //sim::iH+=iHinc;
+			sim::iH=Hfield; //sim::iH+=iHinc;
 
 
 			// Output to screen and file after each field
@@ -168,7 +168,7 @@ int static_hysteresis(){
 
 		// Increment of parity
       iparity+=2;
-      sim::parity=int64_t(iparity);
+      sim::parity=iparity;
 
 	} // End of parity loop
